requirements.txt
# Data handling
pandas==2.1.4
numpy==1.26.2
kagglehub==0.2.0
kaggle==1.6.6

# ML libraries
scikit-learn==1.3.2
xgboost==2.0.3

# MLOps tools
mlflow==2.9.2
dvc==3.38.1

# API
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3
httpx==0.25.2
requests==2.31.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1